Updated sapm- approved

+++ src/algos/wapm/wapm_objects.py
@@ -5,10 +5,10 @@ class WapmObjects():
 
     SYMBOL = tickvuk.parser.get('common', 'symbol')
     TI = int(tickvuk.parser.get(SYMBOL, 'TI'))
-    SL = int(tickvuk.parser.get(SYMBOL, 'SL'))
+    SL = float(tickvuk.parser.get(SYMBOL, 'SL'))
     DTH = float(tickvuk.parser.get(SYMBOL, 'DTH'))
-    TSL = int(tickvuk.parser.get(SYMBOL, 'TSL'))
-    titicks = [] # this list we are going to save time interaval ticks 
+    TSL = float(tickvuk.parser.get(SYMBOL, 'TSL'))
+    titicks = [] # this list we are going to save time interaval ticks
     avgs = []
     LBuy_Position = False # Long buy position
     SSell_Position = False # Short sell position
